[FIX] nullable errors in models [ADD] repository and traits

+++ models/src/submission_ratings.rs
@@ -5,7 +5,7 @@ pub struct SubmissionRating {
     pub id: Uuid,
     pub submission_id: Uuid,
     pub rater_id: Uuid,
-    pub rating_value: u8,
+    pub rating_value: i32,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
